backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
gunicorn>=20.1,<21.0
whitenoise>=6.4,<7.0
setuptools>=69.5.1,<70.0